Update Variance.java
" The average method is being called repeatedly for each element in the array. It reaches same result, same correct answer but cost the computer more energy and memory even."
- Courtesy of "ngarciaiii.github.io"
Basically the coding isn't wrong per say, but Netbeans doesn't like it because it calls the method for each element in the array, and it gives us an error, this fixes that.